requirements.txt
allure-pytest==2.14.2
allure-python-commons==2.14.2
astroid==3.2.4
attrs==25.3.0
autopep8==2.0.4
certifi==2025.4.26
charset-normalizer==3.4.2
dill==0.4.0
docstring-to-markdown==0.16
exceptiongroup==1.3.0
flake8==7.1.2
gitdb==4.0.11
GitPython==3.1.41
greenlet==3.1.1
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.1.0
isort==5.13.2
jedi==0.19.2
Jinja2==3.1.6
MarkupSafe==2.1.5
mccabe==0.7.0
numpy==1.24.4
packaging==25.0
parso==0.8.4
pillow==10.4.0
pixelmatch==0.3.0
platformdirs==4.3.6
playwright==1.48.0
pluggy==1.5.0
pycodestyle==2.12.1
pydocstyle==6.3.0
pyee==12.0.0
pyflakes==3.2.0
pylint==3.2.7
pytest==8.3.5
pytest-base-url==2.1.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-playwright==0.5.2
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.12.2
python-slugify==8.0.4
pytoolconfig==1.3.1
requests==2.32.3
rope==1.13.0
smmap==5.0.1
snowballstemmer==2.2.0
text-unidecode==1.3
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.13.2
ujson==5.10.0
urllib3==2.2.3
whatthepatch==1.0.6
yapf==0.43.0
zipp==3.20.2
